Add streaming network wrapper proof-of-concept.
This commit adds a C++ wrapper around an example output network from the
`schednet` tool. It required some changes in the output signature and
structure, and those changes are documented in the example.

The wrapper turns the neural network into a 1x input, 1x output module,
with streaming FIFO interfaces, using the `hls::stream<T>` types.