implement ->deleted{edges,nodes}
The machinery should be in place to handle edge and node deletions now,
but I still need to write the methods to actually perform the delete
operations. This is tricky for nodes because we have to detect and error
if it would leave dangling edges. It is tricky for edges because we have
to make sure the edge id is removed from the by sink/source arrays.